+++ ChangeLog.md
@@ -6,6 +6,8 @@
 
 - Added the ability to `del` from a history object.
   ([#18](https://github.com/davep/bagofstuff/pull/18))
+- Added `cache.CacheManager`.
+  ([#19](https://github.com/davep/bagofstuff/pull/19))
 
 ## v1.0.0